requirements.txt
discord.py==1.3.1
python-dotenv==0.11.0